改进：添加了FLASH读写数据，USART接收指令功能，数据传输改为发射结束后进行
beta 0.201

+++ Basic/dma/bsp_usart_dma.c
@@ -1,7 +1,7 @@
 
 #include "bsp_usart_dma.h"
 
-extern u8 USART1_Queue[SENDBUFF_SIZE];
+extern float USART1_Queue[SENDBUFF_SIZE];
 	
 /**
   * @brief  USART GPIO 配置,工作参数配置
